Changed Teleop to use finite state machines and encapsulated hardware classes
