--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/stateHardware/Drivetrain.java
@@ -1,51 +1,0 @@
-package org.firstinspires.ftc.teamcode.hardware.stateHardware;
-
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorEx;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.HardwareMap;
-
-public class Drivetrain {
-
-    public DcMotorEx LF;
-    public DcMotorEx LB;
-    public DcMotorEx RF;
-    public DcMotorEx RB;
-
-    HardwareMap hardwareMap;
-
-    public void init(HardwareMap Hmap){
-
-        hardwareMap = Hmap;
-
-        LF = hardwareMap.get(DcMotorEx.class, "LF");
-        LB = hardwareMap.get(DcMotorEx.class, "LB");
-        RF = hardwareMap.get(DcMotorEx.class, "RF");
-        RB = hardwareMap.get(DcMotorEx.class, "RB");
-
-        RF.setDirection(DcMotorSimple.Direction.FORWARD);
-        LF.setDirection(DcMotorSimple.Direction.REVERSE);
-        RB.setDirection(DcMotorSimple.Direction.FORWARD);
-        LB.setDirection(DcMotorSimple.Direction.REVERSE);
-
-        RF.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        RB.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        LF.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-        LB.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
-
-        RF.setPower(0);
-        LF.setPower(0);
-        RB.setPower(0);
-        LB.setPower(0);
-
-    }
-
-    public void setAllPower(double power){
-        RF.setPower(power);
-        LF.setPower(power);
-        RB.setPower(power);
-        LB.setPower(power);
-    }
-
-
-}
